+++ examples/summation_by_lookup_table_with_backend/client/src/main.rs
@@ -118,20 +118,21 @@ async fn send_requests(
                 service_name.clone(),
                 method_name.clone(),
             ))
-            .await
-            .unwrap()
-            .into_inner();
-        // If no status is set, the assumption is no errors occurred
-        if response.status.is_some() {
-            match response.status.as_ref().unwrap().code {
-                0 => println!("Status was explicitly set OK"),
-                other => {
-                    println!("raw response: {response:?}");
-                    anyhow::bail!("RPC error, Status was {other:?}");
-                }
+            .await;
+
+        let inner_response = match response {
+            Ok(inner_response) => {
+                println!("grpc Status was OK");
+                inner_response
             }
-        }
-        let public_output_bytes = prost::bytes::Bytes::from(response.public_output);
+            Err(e) => {
+                println!("rpc failed {e:?}");
+                anyhow::bail!("RPC error, Status was not okay");
+            }
+        };
+
+        let inner_response = inner_response.into_inner();
+        let public_output_bytes = prost::bytes::Bytes::from(inner_response.public_output);
         let parsed_response = IntegerSequenceResponse::decode(public_output_bytes).unwrap();
         println!("Parsed response {task_number}#{i}: {parsed_response:?}");
         assert_eq!(parsed_response.sequence_sum, get_expected_sum_result(sum_start, sum_end));
@@ -170,16 +171,6 @@ async fn send_requests_stream(
     });
     let mut i = 0;
     while let Some(response) = response_stream.message().await.unwrap() {
-        // If no status is set, the assumption is no errors occurred
-        if response.status.is_some() {
-            match response.status.as_ref().unwrap().code {
-                0 => println!("Status was explicitly set OK"),
-                other => {
-                    println!("raw response: {response:?}");
-                    anyhow::bail!("RPC error, Status was {other:?}");
-                }
-            }
-        }
         let public_output_bytes = prost::bytes::Bytes::from(response.public_output);
         let parsed_response = IntegerSequenceResponse::decode(public_output_bytes).unwrap();
         println!("Parsed response {task_number}#{i}: {parsed_response:?}");