+++ .devkit/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 2.11.0 (2026-02-19)
+
+
+### Dependencies
+
+* **deps:** Upgrade gemini-cli to 0.29.2
+
+## 2.10.0 (2026-02-12)
+
+
+### Dependencies
+
+* **deps:** Upgrade gemini-cli to 0.27.3
+* **deps:** Upgrade gemini-cli to 0.28.2
+
 ## 2.9.0 (2026-02-05)