+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.10.0 (2026-03-04)
+
+
+### Dependencies
+
+* **deps:** Bump googletest to v1.17.0
+* **deps:** Bump grpc to v1.78.0
+* **deps:** Bump protobuf to v31.1
+* **deps:** Update abseil-cpp to v20260107.0
+* **deps:** Update DevKit to release-2.11.0
+* Add jinja equivalents to all templates in tools/plugin/tmpl/*
+* Add jinja2 and pip to MODULE.bazel
+
+
+### Features
+
+* Migrate jinja2 template functionality
+* Remove buf-format from EZ SDK
+* Remove EzStatus from SDK repo
+* Remove helloworld_edition
+* Update comment in .pre-commit-config.yaml
+* Use operatorDomain instead of serviceDomain
+
 ## 0.9.0 (2026-02-09)